+++ packages/application/applications/create.js
@@ -1,4 +1,4 @@
-import { ValidationError } from "../entities/ValidationError.js";
+import { ValidationError } from "../errors/ValidationError.js";
 import { validatePhone } from "../outsiders/validators.js";
 import { validateDate } from "../shared/validators.js";
 import { createApplication } from "@reuc/domain/application/createApplication.js";
