+++ packages/application/errors/Warning.js
